wip: move eager/lazy implementation details to base class file, inheriting base init implicitly (_smiles, _table).
